Adding Global Exception Handler and Registrarion Pet
- Fixes to the database
- Adding custom exception handler with @RestControllerAdvice
- Get Pet By Id and Registration of pets.

+++ pets-network-api/src/main/java/com/devandre/petsnetwork/common/BaseEntity.java
@@ -30,9 +30,9 @@ public class BaseEntity {
 
     @CreatedBy
     @Column(nullable = false, updatable = false)
-    private Integer createdBy;
+    private Long createdBy;
 
     @LastModifiedBy
     @Column(insertable = false)
-    private Integer lastModifiedBy;
+    private Long lastModifiedBy;
 }
